Moved cli to teos/cli and client to contrib/client; using relative imports for cli and client
