+++ contrib/client/DEPENDENCIES.md
