+++ contrib/client/exceptions.py
