+++ contrib/client/template.conf
