+++ teos/cli/DEPENDENCIES.md
