+++ teos/cli/INSTALL.md
