+++ teos/cli/template.conf
